Separated DidKey class from fetch_status.py to make it available to plug-ins.
