+++ fetch-ledger-tx/DidKey.py
@@ -1,0 +1,23 @@
+import base58
+import base64
+import nacl.signing
+from indy_vdr.ledger import Request
+
+class DidKey:
+    def __init__(self, seed):
+        seed = seed_as_bytes(seed)
+        self.sk = nacl.signing.SigningKey(seed)
+        self.vk = bytes(self.sk.verify_key)
+        self.did = base58.b58encode(self.vk[:16]).decode("ascii")
+        self.verkey = base58.b58encode(self.vk).decode("ascii")
+
+    def sign_request(self, req: Request):
+        signed = self.sk.sign(req.signature_input)
+        req.set_signature(signed.signature)
+
+def seed_as_bytes(seed):
+    if not seed or isinstance(seed, bytes):
+        return seed
+    if len(seed) != 32:
+        return base64.b64decode(seed)
+    return seed.encode("ascii")
